q_thinking/requirements.txt
google-genai
python-dotenv
matplotlib